telegram-automation/requirements.txt
telethon==1.28.5
python-dotenv==1.0.1
aiofiles==23.2.1
aiosqlite==0.20.0
